Refactor Address Book functions to improve encapsulation
* Remove Some Address Book functions to place them in TutorialTeam and
TutorialClass to maintain good OOP practice.
