+++ src/main/java/seedu/address/logic/commands/allocatestudenttoteamcommands/AllocateStudentToTeamByStuIdCommand.java
@@ -65,7 +65,7 @@ public class AllocateStudentToTeamByStuIdCommand extends AllocateStudentToTeamCo
         TutorialClass tutClass = model.findTutorialClassFromList(tutorialClass, module);
 
         Person student = model.getUniquePersonList().getPerson(studentId);
-        TutorialTeam tutTeam = model.getTutorialTeam(tutClass, tutorialTeam);
+        TutorialTeam tutTeam = tutClass.getTutorialTeam(tutClass, tutorialTeam);
 
         if (student == null) {
             throw new CommandException(MESSAGE_STUDENT_DOES_NOT_EXIST);
@@ -76,7 +76,7 @@ public class AllocateStudentToTeamByStuIdCommand extends AllocateStudentToTeamCo
         }
 
         // throws commandException if any condition fails
-        checkAllocateCondition(model, student, tutClass, tutTeam);
+        checkAllocateCondition(student, tutClass, tutTeam);
         model.allocateStudentToTeam(student, tutTeam);
 
         return new CommandResult(String.format(MESSAGE_SUCCESS, tutTeam));
